Added ENAColor.
