+++ src/xcode/ENA/ENA/Resources/Assets/ena-colors.xcassets/Brand Colors/ENA Brand Blue Color.colorset/Contents.json
