+++ src/xcode/ENA/ENA/Resources/Assets/ena-colors.xcassets/Brand Colors/ENA Brand Burgundy Color.colorset/Contents.json
