+++ src/xcode/ENA/ENA/Resources/Assets/ena-colors.xcassets/Brand Colors/ENA Brand Red Color.colorset/Contents.json
